Add promote_depth for promoting one series to at least the depth of another.
Need this for proper division if the numerator has more terms than the denominator. This is a bit of a fly in the ointment, since it means a/b will give a different answer than a*(1/b) if b has fewer terms than a. This would argue in favor of lazy infinite series (or in favor of not allowing a combination of series of different lenght at all).