Benchmark against fixed size dual type.
The Series type is essentially a linked list. I wanted to check the performance of this implementation with the performance of using simple immutable pair type, so I added a stub type in Dual.jl. The Dual.jl type reaches the 2x overhead that I expected theoretically--it's beating the linked list type by a factor of about 10x.